Correção de tipos de variáveis

+++ CartasSuperTrunfo.c
@@ -11,12 +11,11 @@ int main() {
     // Exemplos de atributos: código da cidade, nome, população, área, PIB, número de pontos turísticos.
     char codigoCidade[4];
     char nomeCidade[20];
-    float populacao;
+    int populacao;
     float area;
-    double PIB;
+    float PIB;
     int qtdPontosTuristicos;
 
-    
     // Cadastro das Cartas:
     // Sugestão: Utilize a função scanf para capturar as entradas do usuário para cada atributo.
     // Solicite ao usuário que insira as informações de cada cidade, como o código, nome, população, área, etc.
@@ -27,23 +26,23 @@ int main() {
     fgets(nomeCidade, sizeof(nomeCidade), stdin);
     fflush(stdin);
     printf("Digite a população da cidade: ");
-    scanf("%f", &populacao);
+    scanf("%d", &populacao);
     printf("Digite a area da cidade: ");
     scanf("%f", &area);
     printf("Digite o PIB da cidade: ");
-    scanf("%lf", &PIB);
+    scanf("%f", &PIB);
     printf("Digite a quantidade de pontos turisticos: ");
     scanf("%d", &qtdPontosTuristicos);
-        
+    
     // Exibição dos Dados das Cartas:
     // Sugestão: Utilize a função printf para exibir as informações das cartas cadastradas de forma clara e organizada.
     // Exiba os valores inseridos para cada atributo da cidade, um por linha.
     printf("\nDados da Carta:\n");
     printf("Código: %s\n", codigoCidade);
     printf("Nome da Cidade: %s", nomeCidade);
-    printf("População: %.2f habitantes\n", populacao);
+    printf("População: %d habitantes\n", populacao);
     printf("Area: %.2f km²\n", area);
-    printf("PIB: $%.2lf\n", PIB);
+    printf("PIB: $%.2f\n", PIB);
     printf("Quantidade de Pontos Turisticos: %d\n", qtdPontosTuristicos);
 
     return 0;
